component reorganization
