+++ ZZZ-Optimizer/components/OSWindow-Variants/setSelectorWindow.tsx
